# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_dataset_api_v1_datasets_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_dataset_api_v1_datasets_upload_post",
        "type": "object"
      },
...
      "CleaningRequest": {
        "properties": {
          "label": {
            "default": "Cleaned revision",
            "maxLength": 160,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/CleaningStep"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "steps"
        ],
        "title": "CleaningRequest",
        "type": "object"
      },
      "CleaningResultStep": {
        "properties": {
          "action": {
            "enum": [
              "drop_duplicates",
              "drop_missing",
              "fill_missing",
              "flag_outliers"
            ],
            "title": "Action",
            "type": "string"
          },
          "column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "rows_after": {
            "title": "Rows After",
            "type": "integer"
          },
          "rows_before": {
            "title": "Rows Before",
            "type": "integer"
          },
          "strategy": {
            "anyOf": [
              {
                "enum": [
                  "mean",
                  "median",
                  "mode",
                  "value"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "required": [
          "action",
          "rows_before",
          "rows_after"
        ],
        "title": "CleaningResultStep",
        "type": "object"
      },
      "CleaningStep": {
        "properties": {
          "action": {
            "enum": [
              "drop_duplicates",
              "drop_missing",
              "fill_missing",
              "flag_outliers"
            ],
            "title": "Action",
            "type": "string"
          },
          "column": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Column"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "strategy": {
            "anyOf": [
              {
                "enum": [
                  "mean",
                  "median",
                  "mode",
                  "value"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "action"
        ],
        "title": "CleaningStep",
        "type": "object"
      },
      "ColumnProfile": {
        "properties": {
          "kind": {
            "enum": [
              "numeric",
              "categorical",
              "boolean",
              "datetime",
              "text"
            ],
            "title": "Kind",
            "type": "string"
          },
          "missing_count": {
            "title": "Missing Count",
            "type": "integer"
          },
          "missing_ratio": {
            "title": "Missing Ratio",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DateRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NumericStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "top_values": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TopValue"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top Values"
          },
          "unique_count": {
            "title": "Unique Count",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "kind",
          "missing_count",
          "missing_ratio",
          "unique_count"
        ],
        "title": "ColumnProfile",
        "type": "object"
      },
...
      "DatasetDetail": {
        "properties": {
          "active_revision": {
            "title": "Active Revision",
            "type": "integer"
          },
          "column_count": {
            "title": "Column Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "file_type": {
            "title": "File Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "preview": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Preview",
            "type": "array"
          },
          "profile": {
            "$ref": "#/components/schemas/DatasetProfile"
          },
          "quality_score": {
            "title": "Quality Score",
            "type": "number"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "transformations": {
            "items": {
              "$ref": "#/components/schemas/CleaningResultStep"
            },
            "title": "Transformations",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "source_filename",
          "file_type",
          "row_count",
          "column_count",
          "active_revision",
          "quality_score",
          "created_at",
          "updated_at",
          "profile",
          "preview",
          "transformations"
        ],
        "title": "DatasetDetail",
        "type": "object"
      },
      "DatasetProfile": {
        "properties": {
          "column_count": {
            "title": "Column Count",
            "type": "integer"
          },
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnProfile"
            },
            "title": "Columns",
            "type": "array"
          },
          "duplicate_rows": {
            "title": "Duplicate Rows",
            "type": "integer"
          },
          "mapping": {
            "$ref": "#/components/schemas/ParticleMapping"
          },
          "missing_cells": {
            "title": "Missing Cells",
            "type": "integer"
          },
          "quality_score": {
            "title": "Quality Score",
            "type": "number"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          }
        },
        "required": [
          "row_count",
          "column_count",
          "missing_cells",
          "duplicate_rows",
          "quality_score",
          "columns",
          "mapping"
        ],
        "title": "DatasetProfile",
        "type": "object"
      },
...
      "DateRange": {
        "properties": {
          "end": {
            "title": "End",
            "type": "string"
          },
          "start": {
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "DateRange",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NumericStats": {
        "properties": {
          "max": {
            "title": "Max",
            "type": "number"
          },
          "mean": {
            "title": "Mean",
            "type": "number"
          },
          "median": {
            "title": "Median",
            "type": "number"
          },
          "min": {
            "title": "Min",
            "type": "number"
          },
          "outlier_count": {
            "title": "Outlier Count",
            "type": "integer"
          },
          "q1": {
            "title": "Q1",
            "type": "number"
          },
          "q3": {
            "title": "Q3",
            "type": "number"
          },
          "std": {
            "title": "Std",
            "type": "number"
          }
        },
        "required": [
          "mean",
          "median",
          "std",
          "min",
          "max",
          "q1",
          "q3",
          "outlier_count"
        ],
        "title": "NumericStats",
        "type": "object"
      },
      "ParticleMapping": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time"
          },
          "x": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "z": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z"
          }
        },
        "title": "ParticleMapping",
        "type": "object"
      },
      "ParticlePoint": {
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          },
          "z": {
            "title": "Z",
            "type": "number"
          }
        },
        "required": [
          "id",
          "x",
          "y",
          "z",
          "color",
          "label",
          "values"
        ],
        "title": "ParticlePoint",
        "type": "object"
      },
      "ParticleRequest": {
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "limit": {
            "default": 1000,
            "maximum": 5000.0,
            "minimum": 50.0,
            "title": "Limit",
            "type": "integer"
          },
          "x": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          },
          "z": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Z"
          }
        },
        "title": "ParticleRequest",
        "type": "object"
      },
      "ParticleResponse": {
        "properties": {
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "mapping": {
            "$ref": "#/components/schemas/ParticleMapping"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ParticlePoint"
            },
            "title": "Points",
            "type": "array"
          }
        },
        "required": [
          "dataset_id",
          "mapping",
          "points"
        ],
        "title": "ParticleResponse",
        "type": "object"
      },
      "RowPage": {
        "properties": {
          "columns": {
            "items": {
              "type": "string"
            },
            "title": "Columns",
            "type": "array"
          },
          "dataset_id": {
            "title": "Dataset Id",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "dataset_id",
          "revision",
          "offset",
          "limit",
          "total",
          "columns",
          "rows"
        ],
        "title": "RowPage",
        "type": "object"
      },
...
      "TopValue": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "ratio": {
            "title": "Ratio",
            "type": "number"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "count",
          "ratio"
        ],
        "title": "TopValue",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Vibe Data Universe",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/datasets/upload": {
      "post": {
        "operationId": "upload_dataset_api_v1_datasets_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_dataset_api_v1_datasets_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}": {
      "delete": {
        "operationId": "delete_dataset_api_v1_datasets__dataset_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Dataset",
        "tags": [
          "datasets"
        ]
      },
      "get": {
        "operationId": "dataset_detail_api_v1_datasets__dataset_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Detail",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/clean": {
      "post": {
        "operationId": "clean_dataset_api_v1_datasets__dataset_id__clean_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CleaningRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clean Dataset",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/particles": {
      "post": {
        "operationId": "dataset_particles_api_v1_datasets__dataset_id__particles_post",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Particles",
        "tags": [
          "datasets"
        ]
      }
    },
    "/api/v1/datasets/{dataset_id}/rows": {
      "get": {
        "operationId": "dataset_rows_api_v1_datasets__dataset_id__rows_get",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dataset Rows",
        "tags": [
          "datasets"
        ]
      }
    },
...
  }
}